feat(embed): add audio and video multimodal embedding support (#1)
Add audio and video embedding support to the multimodal embedding system.

## Summary
- Add AudioEmbedInput, VideoEmbedInput, MultimodalInput types
- Add MediaModality enum (Text, Image, Audio, Video)
- Extend Embedder trait with embed_audio(), embed_video(), embed_unified()
- Add supported_modalities() for capability discovery
- HTTP backend supports audio/video with MIME detection
- Xinference backend gracefully reports text-only support
- Model registry extended with audio (CLAP, MERT), video (InternVideo2), omni (BLIP2, ImageBind) models
- Update sidecar/models.json with new model definitions
- 99 tests passing

Implements Phase 3 of multimodal embedding roadmap.

+++ src/store/mod.rs
@@ -647,12 +647,12 @@ fn extract_vector_sizes(info: &GetCollectionInfoResponse) -> Option<Vec<(String,
 
     match config {
         qdrant_client::qdrant::vectors_config::Config::Params(params) => {
-            Some(vec![("default".to_string(), params.size as u64)])
+            Some(vec![("default".to_string(), params.size)])
         }
         qdrant_client::qdrant::vectors_config::Config::ParamsMap(map) => {
             let mut sizes = Vec::new();
             for (name, params) in &map.map {
-                sizes.push((name.clone(), params.size as u64));
+                sizes.push((name.clone(), params.size));
             }
             if sizes.is_empty() {
                 None
